migração para nova padronização de pastas do modulo l10n_br_stock_account
