+++ l10n_br_stock_account/__openerp__.py
@@ -1,20 +1,20 @@
-# -*- encoding: utf-8 -*-
+# -*- coding: utf-8 -*-
 ###############################################################################
-#                                                                             #
-# Copyright (C) 2014  Renato Lima - Akretion                                  #
-#                                                                             #
-#This program is free software: you can redistribute it and/or modify         #
-#it under the terms of the GNU Affero General Public License as published by  #
-#the Free Software Foundation, either version 3 of the License, or            #
-#(at your option) any later version.                                          #
-#                                                                             #
-#This program is distributed in the hope that it will be useful,              #
-#but WITHOUT ANY WARRANTY; without even the implied warranty of               #
-#MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the                #
-#GNU Affero General Public License for more details.                          #
-#                                                                             #
-#You should have received a copy of the GNU Affero General Public License     #
-#along with this program.  If not, see <http://www.gnu.org/licenses/>.        #
+#
+# Copyright (C) 2014  Renato Lima - Akretion
+#
+#This program is free software: you can redistribute it and/or modify
+#it under the terms of the GNU Affero General Public License as published by
+#the Free Software Foundation, either version 3 of the License, or
+#(at your option) any later version.
+#
+#This program is distributed in the hope that it will be useful,
+#but WITHOUT ANY WARRANTY; without even the implied warranty of
+#MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#GNU Affero General Public License for more details.
+#
+#You should have received a copy of the GNU Affero General Public License
+#along with this program.  If not, see <http://www.gnu.org/licenses/>.
 ###############################################################################
 
 {
@@ -31,9 +31,9 @@
         'account_fiscal_position_rule_stock',
     ],
     'data': [
-        'l10n_br_stock_account_data.xml',
-        'stock_account_view.xml',
-        'res_company_view.xml',
+        'data/l10n_br_stock_account_data.xml',
+        'views/stock_account_view.xml',
+        'views/res_company_view.xml',
         'wizard/stock_invoice_onshipping_view.xml',
         'security/ir.model.access.csv',
     ],
