+++ l10n_br_stock_account/models/__init__.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 ###############################################################################
 #
-# Copyright (C) 2014  Renato Lima - Akretion
+# Copyright (C) 2015  Renato Lima - Akretion
 #
 #This program is free software: you can redistribute it and/or modify
 #it under the terms of the GNU Affero General Public License as published by
@@ -16,6 +16,3 @@
 #You should have received a copy of the GNU Affero General Public License
 #along with this program.  If not, see <http://www.gnu.org/licenses/>.
 ###############################################################################
-
-from . import models
-from . import wizard
