+++ l10n_br_stock_account/views/res_company_view.xml
