+++ l10n_br_stock_account/views/stock_account_view.xml
